Refactor all Java files to resolve Checkstyle violations
